+++ typescript/ui/src/core/components/Layout.tsx
@@ -4,7 +4,9 @@ import { useHistoryContext } from "../contexts/HistoryContext/useHistoryContext"
 import { useThemeContext } from "../contexts/ThemeContext/useThemeContext";
 import { Theme } from "../contexts/ThemeContext/ThemeContext";
 import DelayedTooltip from "./DelayedTooltip";
-import { ReactNode } from "react";
+import { ReactNode, useState } from "react";
+import { useConfigContext } from "../contexts/ConfigContext/useConfigContext";
+import DebugPanel from "./DebugPanel";
 
 interface LayoutProps {
   children: ReactNode;
@@ -26,8 +28,10 @@ export default function Layout({
   logo,
   classNames,
 }: LayoutProps) {
+  const { config } = useConfigContext();
   const { clearHistory, stopAnswering } = useHistoryContext();
   const { setTheme, theme } = useThemeContext();
+  const [isDebugOpened, setDebugOpened] = useState(false);
 
   const toggleTheme = () => {
     setTheme(theme === Theme.DARK ? Theme.LIGHT : Theme.DARK);
@@ -71,7 +75,7 @@ export default function Layout({
   }
 
   return (
-    <div className={cn("flex h-full min-h-[48rem] justify-center py-4")}>
+    <div className="flex h-full min-h-[48rem] justify-center py-4">
       <div className="flex w-full flex-col px-4 sm:max-w-[1200px]">
         <header
           className={cn(
@@ -131,6 +135,21 @@ export default function Layout({
                 )}
               </Button>
             </DelayedTooltip>
+            {config.debug_mode && (
+              <DelayedTooltip content="Toggle debug panel" placement="bottom">
+                <Button
+                  isIconOnly
+                  aria-label={`${isDebugOpened ? "Open" : "Close"} debug panel`}
+                  variant="ghost"
+                  onPress={() => setDebugOpened((o) => !o)}
+                >
+                  <Icon icon="heroicons:bug-ant" />
+                  {isDebugOpened && (
+                    <div className="absolute left-0 right-0 top-1/2 h-0.5 -rotate-45 bg-default-500" />
+                  )}
+                </Button>
+              </DelayedTooltip>
+            )}
           </div>
         </header>
         <main className="flex h-full overflow-hidden">
@@ -144,6 +163,7 @@ export default function Layout({
           </div>
         </main>
       </div>
+      <DebugPanel isOpen={isDebugOpened} />
     </div>
   );
 }